Combat System change from time to frame based
- Changed combat system from time-based to frame-based
- All attacks now have:
1) Start up frames
2) Active frames
3) Recovery frames

+++ game_project/source/MathFunctions.cpp
@@ -2,7 +2,7 @@
 #include <AEEngine.h>
 
 namespace Vectors {
-	f32 magnitude(f32 vecX, f32 vecY) {
+	double magnitude(f32 vecX, f32 vecY) {
 		return sqrt(vecX * vecX + vecY * vecY);
 	}
